# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Herstory AI - Backend REST API Specification",
    "description": "Herstory AI 백엔드 시스템 OpenAPI 3.0 명세서입니다.\n- Auth API: 로그인 & VIP 허브\n- Journey API: OCR 스캔, 기후 분석, Live Card 위젯, Apple Wallet .pkpass\n- Cart API: 스마트 장바구니 & ChoiceFit 피팅 신청\n- Store API: 면세점 체크인 & 매장 직원 태블릿 SSE 실시간 알림\n- Order API: 면세 결제 & Herstory Miles 적립\n- Care API: 현지 럭셔리 부티크 스팟, Google Maps, OpenAI 가죽 케어, 시티 스탬프, FCM\n- Flight API: 항공편 실시간 운항 정보 조회\n- Health API: 시스템 헬스 모니터링",
    "version": "v1.0.0",
    "contact": {
      "name": "Herstory Tech Team",
      "email": "tech@herstory.com"
    }
  },
  "servers": [
    {
      "url": "https://herstory-backend.onrender.com",
      "description": "Render Production Server (Live)"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/v1/journey/analysis/{journeyId}": {
      "get": {
        "tags": ["2. Journey API (여정 & 큐레이션)"],
        "summary": "목적지 기후 및 여행 분석 데이터 조회",
        "description": "목적지 날씨 및 기후 정보를 분석하여 맞춤형 럭셔리 상품을 추천합니다.",
        "operationId": "analyzeJourney",
        "parameters": [
          {
            "name": "journeyId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyAnalysisResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/care/visetos-spots": {
      "get": {
        "tags": ["6. Care API (현지 부티크 스팟 & 케어)"],
        "summary": "현지 글로벌 럭셔리 매장 스팟 탐색 및 가죽 케어 메시지 조회",
        "description": "목적지의 현지 럭셔리 플래그십 스토어 및 Care Desk 정보를 조회하고 맞춤형 케어 가이드를 반환합니다.",
        "operationId": "getVisetosSpots",
        "parameters": [
          {
            "name": "memberId",
            "in": "query",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CareResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProductCategory": {
        "type": "string",
        "enum": ["WATERPROOF", "BACKPACK", "TRAVEL_BAG", "ACCESSORY", "LEATHER_CARE"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "integer", "example": 400 },
          "error": { "type": "string", "example": "Bad Request" },
          "message": { "type": "string", "example": "해당 회원을 찾을 수 없습니다." },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
...
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "example": 1 },
          "name": { "type": "string", "example": "럭셔리 클래식 방수 스타크 백팩" },
          "category": { "$ref": "#/components/schemas/ProductCategory" },
          "price": { "type": "number", "example": 1250000.00 },
          "stock": { "type": "integer", "example": 45 },
          "imageUrl": { "type": "string", "example": "https://images.unsplash.com/photo-1553062407-98eeb64c6a62?auto=format&fit=crop&w=800&q=80" },
          "description": { "type": "string", "example": "열대 스콜 방수 코팅 캔버스 백팩" },
          "isVipExclusive": { "type": "boolean", "example": false }
        }
      },
      "JourneyAnalysisResponse": {
        "type": "object",
        "properties": {
          "journeyId": { "type": "integer", "format": "int64", "example": 1 },
          "destination": { "type": "string", "example": "BKK (방콕 수완나품)" },
          "weatherInfo": { "type": "string", "example": "기온 32.5°C, 습도 88%" },
          "climateSummary": { "type": "string", "example": "고온 다습한 열대 몬순 기후" },
          "recommendationReason": { "type": "string", "example": "방수 에센셜 라인업 제안" },
          "recommendedProducts": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Product" }
          }
        }
      },
...
      "VisetosSpot": {
        "type": "object",
        "properties": {
          "spotName": { "type": "string", "example": "Herstory Siam Paragon Flagship" },
          "address": { "type": "string", "example": "991 Rama I Rd, Bangkok" },
          "locationType": { "type": "string", "example": "Herstory Flagship & VIP Lounge" },
          "latitude": { "type": "number", "example": 13.7466 },
          "longitude": { "type": "number", "example": 100.5349 },
          "careServiceAvailable": { "type": "string", "example": "가죽 방수 코팅 케어 제공" }
        }
      },
      "CareResponse": {
        "type": "object",
        "properties": {
          "destination": { "type": "string", "example": "Bangkok" },
          "pushNotificationMessage": { "type": "string", "example": "방콕 도착 환영 가이드" },
          "visetosSpots": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/VisetosSpot" }
          }
        }
      },
...
    }
  }
}